+++ happiness_service/src/main/java/klosterteam/hibernate/Roles.java
@@ -17,20 +17,21 @@ import javax.persistence.*;
 public class Roles implements Serializable {
     @Id
     @Column(name="id", unique = true, nullable = false)
-    private long id;
+    private short id;
     @Column(name="role", unique = true, nullable = false, length = 16)
     private String role;
 
     public Roles() {
     }
 
-    public long getId() {
+    public short getId() {
         return id;
     }
 
-    public void setId(long id) {
+    public void setId(short id) {
         this.id = id;
     }
+
 
     public String getRole() {
         return role;
